+++ tgbdual_desktop/src/emulator_time.h
@@ -4,9 +4,9 @@
 
 class emulator_time {
 public:
-    using time_provider = std::function<long()>;
-    using sleep_provider = std::function<void(long)>;
-    static long current_time();
+    using time_provider = std::function<uint32_t()>;
+    using sleep_provider = std::function<void(uint32_t)>;
+    static uint32_t current_time();
     static void set_time_provider(time_provider provider);
     static void set_sleep_provider(sleep_provider provider);
     static void sleep(uint32_t duration);
